feat: RDP, 평균 고도 계산 로직을 적용한 러닝/코스 저장 API 고도화

+++ src/main/java/soma/ghostrunner/domain/running/application/dto/TelemetryDto.java
@@ -14,7 +14,7 @@ public class TelemetryDto {
     private Double lng;
     private Double dist;
     private Double pace;
-    private Integer alt;
+    private Double alt;
     private Integer cadence;
     private Integer bpm;
     private Boolean isRunning;
